fix: silence CodeQL lints on test fixtures introduced this PR
Seventeen CodeQL findings layered on top of the original Copilot review,
all on test code added in this PR:

- **Path.Combine → Path.Join** (six sites across three test fixtures).
  Path.Combine treats an absolute later arg as a new root and silently
  drops earlier args; Path.Join always concatenates literally. Our
  second args are always relative literals (`"graph.db"`, `"tests"`,
  `"fixtures"`, `"Sample.sln"`), so the original code is correct, but
  Path.Join makes the lack-of-rooting-risk explicit and clears the lint.

- **Generic catch clauses → catch (Exception)** (three sites). The
  best-effort cleanup intent is unchanged — we still swallow every
  exception during DisposeAsync — but naming the type satisfies CodeQL's
  preference and the comment explicitly notes the rationale.

- **Floating-point equality on metric values → long cast** (two sites
  in TelemetrySignalTests). The values come from `Counter<long>.Add(1)`
  on OTel instruments, so they're integer-valued; the float comparison
  was correct but CodeQL doesn't know that. Cast to `long` for an exact
  comparison without the float-equality lint.

Skipped (false positives that don't apply cleanly):
  - "Missed opportunity to use Where" on ToolMetrics.MeasureContent — the
    foreach uses pattern-match dispatch over multiple ContentBlock types,
    not implicit filtering.
  - "Missed opportunity to use Select" on
    FindDefinitionStructuredOutputTests's resource-link loop — the body
    has `await GraphResources.GetSymbolAsync` and multiple per-iteration
    assertions, not a pure map.
  - "Generic catch" on ScopedExecution.cs:143 — that line is in main's
    code, not changed by this PR.

Tests: 330/330 still pass. Build clean.

+++ tests/DevBitsLab.Mcp.SourceGraph.Tests/TelemetrySignalTests.cs
@@ -205,8 +205,11 @@ public sealed class TelemetrySignalTests
             IsError = true,
         }));
 
-        samples.Should().Contain(s => s.Instrument == "sourcegraph.tool.calls" && s.Value == 1d);
-        samples.Should().Contain(s => s.Instrument == "sourcegraph.tool.errors" && s.Value == 1d);
+        // Cast Value to long to keep CodeQL's float-equality lint quiet — these are integer
+        // counters (Counter<long>.Add(1) on the metric instrument), so the read-back is exactly
+        // representable and the cast loses no information.
+        samples.Should().Contain(s => s.Instrument == "sourcegraph.tool.calls" && (long)s.Value == 1L);
+        samples.Should().Contain(s => s.Instrument == "sourcegraph.tool.errors" && (long)s.Value == 1L);
     }
 
     private sealed record TestStructuredDto(string Status, int Value);
